+++ src/index.ts
@@ -368,6 +368,8 @@ class TheAlgorithm {
         MastoApi.instance.setSemaphoreConcurrency(Config.api.maxConcurrentRequestsInitial);
         this.dataPoller && clearInterval(this.dataPoller!);
         this.dataPoller = undefined;
+        this.cacheUpdater && clearInterval(this.cacheUpdater!);
+        this.cacheUpdater = undefined;
         this.hasProvidedAnyTootsToClient = false;
         this.loadingStatus = READY_TO_LOAD_MSG;
         this.loadStartedAt = null;
@@ -655,7 +657,9 @@ class TheAlgorithm {
         if (this.isLoading()) return;
         const newTotalNumTimesShown = this.feed.reduce((sum, toot) => sum + (toot.numTimesShown ?? 0), 0);
         if (this.totalNumTimesShown == newTotalNumTimesShown) return;
-        console.debug(`[updateTootCache()] saving ${this.feed.length} toots with ${newTotalNumTimesShown} times shown (old: ${this.totalNumTimesShown})`);
+        const numShownToots = this.feed.filter(toot => toot.numTimesShown).length;
+        const msg = `[updateTootCache()] saving ${this.feed.length} toots with ${newTotalNumTimesShown} times shown`;
+        console.debug(`${msg} on ${numShownToots} toots (previous totalNumTimesShown: ${this.totalNumTimesShown})`);
         await Storage.set(CacheKey.TIMELINE, this.feed);
         this.totalNumTimesShown = newTotalNumTimesShown;
     }